fix pre-release checks in constraints
Pre-releases should not be automatically accepted as candidate versions
for constraints without a pre-release component. This follows the logic
documented for RubyGems which appears to be the inspiration for the
constraint operators, and matches the behavior of other version
packages.

Adding a pre-release component to a pessimistic constraint will limit
accepted versions to pre-releases only matching the base segments.
Adding a pre-release to other constraint types will accept any
comparable version, but pre-releases must still match the base segments.