Segments64 makes a copy to match Segments
Fixes #24

Segments and Segments64 should behave identically and its surprising
they don't. I don't think any downstreams depend on the performance
guarantees of returning the exact slice (no copy), especially since the
segments are so small to begin with, but if that's the case we can
always introduce a new func to return the reference.